Switch to price_tier on PrintPrice
Started to realize that I was overcomplicating my print price model when it came to querying. Currently, printer can set arbitrary price tiers, but I'm going to set them, then use that as a value.

Then, instead of allowing quote of colors to be a string field and putting a conditional, I'm going to do a drop down and query off that value.